Fix PR 113 CI: move zod impl out of index.ts + harden coverage script
Two fixes bundled because they address the same red check:

1. packages/zod/src/index.ts had the full implementation, but vitest
   configs across every package exclude src/index.ts from coverage
   (standard: index.ts is just a re-export barrel). Result: zod's
   coverage came back 0/0 everywhere and vitest wrote "Unknown" for
   every pct field. Implementation moves to src/fromZod.ts; index.ts
   becomes a one-liner re-export, so coverage now measures real files
   (90% statements / 75% branches / 90% functions / 92% lines).

2. .github/scripts/coverage-comment.mjs's pct() helper called
   n.toFixed(2) directly — which throws when vitest emits the literal
   string "Unknown" for an empty coverage summary. Harden pct() to
   coerce via Number() and fall back to "—" on non-finite values, so
   one mis-configured package can't take down the whole coverage comment.
